+++ src/server/scripts/Commands/cs_modify.cpp
@@ -994,13 +994,13 @@ public:
         if (handler->HasLowerSecurity(target, 0))
             return false;
 
-        int32 addmoney = atoi((char*)args);
+        int64 addmoney = atoi((char*)args);
 
-        uint32 moneyuser = target->GetMoney();
+        uint64 moneyuser = target->GetMoney();
 
         if (addmoney < 0)
         {
-            int32 newmoney = int32(moneyuser) + addmoney;
+            int64 newmoney = int32(moneyuser) + addmoney;
 
             sLog->outDetail(handler->GetSkyfireString(LANG_CURRENT_MONEY), moneyuser, addmoney, newmoney);
             if (newmoney <= 0)
